llm-training-platform/requirements.txt
# LLM Training Platform Requirements

# Core Dependencies
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic>=1.10.7
sqlalchemy>=2.0.12
psycopg2-binary>=2.9.6
alembic>=1.10.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.1.0
loguru>=0.7.0
python-dotenv>=1.0.0
tenacity>=8.2.2
httpx>=0.24.0
email-validator>=2.0.0

# Security
cryptography>=40.0.2
pyOpenSSL>=23.1.1
certifi>=2023.5.7

# Document Processing
pytesseract>=0.3.10
pdf2image>=1.16.3
python-docx>=0.8.11
pdfminer.six>=20221105
pillow>=9.5.0
opencv-python>=4.7.0.72
langdetect>=1.0.9
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# Vector Database
pymilvus>=2.2.8

# Machine Learning
torch>=2.0.0
transformers>=4.28.1
datasets>=2.12.0
sentence-transformers>=2.2.2
accelerate>=0.19.0
bitsandbytes>=0.38.0
peft>=0.3.0
trl>=0.4.7

# Utilities
tqdm>=4.65.0
pandas>=2.0.1
numpy>=1.24.3
matplotlib>=3.7.1
scikit-learn>=1.2.2
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
